Apply code scanning fix for too few arguments to formatting function
I was thinking it'd try switching to the `%zu` specifier next, but I guess adding a cast could work, too...

+++ src/gdb/tracepoint.c
@@ -1274,7 +1274,7 @@ collect_symbol(struct collection_list *collect,
       if (info_verbose)
 	{
 	  printf_filtered("LOC_REGPARM_ADDR %s: Collect %lu bytes at offset ",
-			  DEPRECATED_SYMBOL_NAME(sym), len);
+			  DEPRECATED_SYMBOL_NAME(sym), (unsigned long)len);
 	  printf_vma(offset);
 	  printf_filtered(" from reg %d\n", reg);
 	}
